+++ CHANGELOG.md
@@ -10,6 +10,8 @@ v1.8.0 (unreleased)
     `config:",interpolate"` to support reading environment variables in them.
     See the `TransportSpec` documentation for more information.
 -   http: Added support for configuring the HTTP transport using x/config.
+-   tchannel: Added support for configuring the TChannel transport using
+    x/config.
 -   Options `thrift.Multiplexed` and `thrift.Enveloped` may now be provided for
     Thrift clients constructed by `yarpc.InjectClients` by adding a `thrift`
     tag to the corresponding struct field with the name of the option. See the